+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version X.Y.Z (YYYY-MM-DD)
 
 - Add support for HTTP and SOCKS proxy servers (#16 - Thanks to @PhMemmel)
+- Fix parsing of boolean / int values during config tree building from environment variables
 
 
 ## Version 2.1.4 (2025-01-02)